Replace files and perform importation refactorings.
